Initial acceralation and movement of the Knight

+++ src/main/java/HollowKnight/view/elements/ParticleViewer.java
@@ -8,6 +8,6 @@ import java.io.IOException;
 public class ParticleViewer implements ElementViewer<Particle>{
     @Override
     public void draw(Particle model, GUI gui, long time) throws IOException {
-        gui.drawPixel(model.getPosition().x(), model.getPosition().y(), model.getColor());
+        gui.drawPixel((int)model.getPosition().x(), (int)model.getPosition().y(), model.getColor());
     }
 }
